添加了DS18B20Example

+++ WJ-C51/DS18B20.h
@@ -8,12 +8,12 @@ void DS18B20_WriteBit(char bitval);
 void DS18B20_WriteByte(uchar dat);
 void DS18B20_SendChangeCmd();
 void DS18B20_SendReadCmd();
-int DS18B20_GetTmpValue();
+float DS18B20_GetTmpValue();
 DS18B20_ReadRomCord(void);
 uchar DS18B20_CRC8();
 void DS18B20_RomChar(char *RomChar);
 void DS18B20_GetTmpStr(char *str);
-void DS18B20_temperToStr(int v, char *str);
+void DS18B20_temperToStr(float v, char *str);
 
 
 #endif
